requirements.txt
airtable==0.4.8
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
appnope==0.1.4
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
attrs==23.1.0
awscli==1.32.61
backoff==2.2.1
blinker==1.7.0
botocore==1.34.61
branca==0.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.37.0
colorama==0.4.4
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
dateutils==0.6.12
debugpy==1.8.1
decorator==5.1.1
distro==1.8.0
docutils==0.16
duckdb==0.10.1
executing==2.0.1
folium==0.15.1
fonttools==4.47.0
gitdb==4.0.11
GitPython==3.1.40
gw_dsl_parser==0.1.48a3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==6.11.0
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kanaries_track==0.0.4
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
monotonic==1.6
nest-asyncio==1.6.0
numpy==1.26.2
openai==1.6.1
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.4
parso==0.8.4
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.2.1
plotly==5.22.0
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.5.1
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
pygwalker==0.4.8.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
rsa==4.7.2
s3transfer==0.10.0
seaborn==0.13.2
segment-analytics-python==2.2.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.30
sqlglot==23.14.0
stack-data==0.6.3
streamlit==1.29.0
streamlit-folium==0.18.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
wasmtime==12.0.0
watchdog==3.0.0
wcwidth==0.2.13
widgetsnbextension==4.0.10
xyzservices==2023.10.1
zipp==3.17.0
